+++ Helpers/Math.php
@@ -4,28 +4,41 @@ namespace is\Helpers;
 
 class Math
 {
+    /**
+     * Функция получения случайного num-значного числа
+     *
+     * @param integer $num
+     * @return void
+     */
     public static function random($num = 4)
     {
-        // функция получения случайного num-значного числа
-
         $a = 1 . Strings::multiply('0', $num - 1);
         $b = Strings::multiply('9', $num);
 
         return rand((int) $a, (int) $b);
     }
 
+    /**
+     * Преобразование в правильные числа
+     *
+     * @param [type] $a
+     * @return void
+     */
     public static function convert($a)
     {
-        // преобразование в правильные числа
         return System::typeTo(Strings::replace($a, ',', '.'), 'numeric');
     }
 
+    /**
+     * Функция разбора числа
+     *
+     * @param [type] $a
+     * @return void
+     */
     public static function fraction($a)
     {
         $a = self::convert($a);
-
         $split = Strings::split($a, '.');
-
         return [
             'original' => $a,
             'int' => $split[0],
@@ -35,10 +48,16 @@ class Math
         ];
     }
 
+    /**
+     * Функция округления чисел
+     *
+     * @param [type] $a
+     * @param integer $precision
+     * @param [type] $mode
+     * @return void
+     */
     public static function precision($a, $precision = 1, $mode = null)
     {
-        // функция округления чисел
-
         if (!$a) {
             return 0;
         }
@@ -53,9 +72,9 @@ class Math
             $result = $precision * ceil($a / $precision);
         } else {
             if (
-                $mode !== 'down' &&
-                $mode !== 'even' &&
-                $mode !== 'odd'
+                $mode !== 'down'
+                && $mode !== 'even'
+                && $mode !== 'odd'
             ) {
                 $mode = 'up';
             }
@@ -66,33 +85,63 @@ class Math
         return $result == '-0' ? 0 : $result;
     }
 
+    /**
+     * Функция точного сложения любых десятичных чисел
+     *
+     * @param [type] $a
+     * @param [type] $b
+     * @return void
+     */
     public static function add($a, $b)
     {
-        // функция точного сложения любых десятичных чисел
-        return self::convert($a) + self::convert($b);
+        return (float) self::convert($a) + (float) self::convert($b);
     }
 
+    /**
+     * Функция точного вычитания любых десятичных чисел
+     *
+     * @param [type] $a
+     * @param [type] $b
+     * @return void
+     */
     public static function sub($a, $b)
     {
-        // функция точного вычитания любых десятичных чисел
-        return self::convert($a) - self::convert($b);
+        return (float) self::convert($a) - (float) self::convert($b);
     }
 
+    /**
+     * Синоним предыдущей функции
+     *
+     * @param [type] $a
+     * @param [type] $b
+     * @return void
+     */
     public static function diff($a, $b)
     {
         return self::sub($a, $b);
     }
 
+    /**
+     * Функция точного умножения любых десятичных чисел
+     *
+     * @param [type] $a
+     * @param [type] $b
+     * @return void
+     */
     public static function multiply($a, $b)
     {
-        // функция точного умножения любых десятичных чисел
-        return self::convert($a) * self::convert($b);
+        return (float) self::convert($a) * (float) self::convert($b);
     }
 
+    /**
+     * Функция точного деления любых десятичных чисел
+     *
+     * @param [type] $a
+     * @param [type] $b
+     * @return void
+     */
     public static function divide($a, $b)
     {
-        // функция точного деления любых десятичных чисел
-
         $a = self::fraction($a);
         $b = self::fraction($b);
